+++ src/main/kotlin/ui/instance/context/modification/InstanceEditDialog.kt
@@ -2,6 +2,7 @@ package dev.nycode.omsilauncher.ui.instance.context.modification
 
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
+import cafe.adriel.lyricist.LocalStrings
 import dev.nycode.omsilauncher.instance.Instance
 
 @Composable
@@ -10,14 +11,21 @@ fun InstanceEditDialog(
     onCloseRequest: () -> Unit,
     saveInstance: (InstanceModificationState) -> Unit,
 ) {
+    val strings = LocalStrings.current
+    val title = if (instance.isBaseInstance) {
+        strings.editBaseInstance
+    } else {
+        strings.editInstanceTitle(instance.name)
+    }
     InstanceDialog(
-        "Instanz ${instance.name} bearbeiten",
-        "Instanz ${instance.name} bearbeiten",
+        title,
+        title,
         instance,
         true,
+        instance.isBaseInstance,
         onCloseRequest = onCloseRequest,
         saveButtonLabel = {
-            Text("Speichern")
+            Text(strings.save)
         },
         onUpdate = saveInstance
     )